refactor: add getPath module
